Compare against sequence name strings instead of IDs
This change looks up / compares sequence name strings instead of their FName comparison indices, as the indices change between builds. The string values are looked up by an FName's comparison index, which is used to compute the address of an FNameEntry, which contains a header and the underlying string data.

Unfortuantely, a build-specific value is still needed (gNameBlocksDebugOffset), as the location in memory of the FName pool changes between builds as well, but it means that only one value needs to be updated for future builds instead of many.

I did some simple testing and have verified this works for all current Windows builds of the game (Steam, Xbox store, and DRM-free); however, I have not tested each split individually.

Hopefully this makes future updates to this script easier!